Emit error for deprecated media features
When the CSS3 profile is selected, this change causes an error to be
emitted for the following deprecated media features:
device-aspect-ratio, device-height, device-width — as well as for their
corresponding min- and max- prefixed versions.

https://www.w3.org/TR/2017/CR-mediaqueries-4-20170905/#mf-deprecated
states:

> The following media features are deprecated. They kept for backward
> compatibility, but are not appropriate for newly written style sheets.
> Authors must not use them.

Otherwise, without this change, no messages (not even warnings) are
reported for the (mis)use of those media features.